Album DTO

+++ musify-backend/src/main/java/com/distributedApplications/Musify/repository/AlbumRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface AlbumRepository extends JpaRepository<Album, Integer> {
+public interface AlbumRepository extends JpaRepository<Album, Long> {
 
     Album findByTitle(String name);
